+++ cmd/dfs/cmd/start.go
@@ -69,9 +69,12 @@ func startHttpService() {
 	userRouter.HandleFunc("/name", handler.SaveUserNameHandler).Methods("POST")
 	userRouter.HandleFunc("/contact", handler.SaveUserContactHandler).Methods("POST")
 
+	userRouter.HandleFunc("/stat", handler.GetUserStatHandler).Methods("GET")
 	userRouter.HandleFunc("/avatar", handler.GetUserAvatarHandler).Methods("GET")
 	userRouter.HandleFunc("/name", handler.GetUserNameHandler).Methods("GET")
 	userRouter.HandleFunc("/contact", handler.GetUserContactHandler).Methods("GET")
+	userRouter.HandleFunc("/share/inbox", handler.GetUserSharingInboxHandler).Methods("GET")
+	userRouter.HandleFunc("/share/outbox", handler.GetUserSharingOutboxHandler).Methods("GET")
 
 	// pod related handlers
 	podRouter := router.PathPrefix("/v0/pod/").Subrouter()
