+++ pkg/user/stat.go
@@ -1,0 +1,36 @@
+/*
+Copyright © 2020 intOS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package user
+
+import "github.com/jmozah/intOS-dfs/pkg/account"
+
+type Stat struct {
+	Name      string `json:"name"`
+	Reference string `json:"reference"`
+}
+
+func (u *Users) GetUserStat(userInfo *Info) (*Stat, error) {
+	if !u.IsUsernameAvailable(userInfo.name, u.dataDir) {
+		return nil, ErrInvalidUserName
+	}
+
+	stat := &Stat{
+		Name:      userInfo.name,
+		Reference: userInfo.GetAccount().GetAddress(account.UserAccountIndex).Hex(),
+	}
+	return stat, nil
+}
